refactor: 농부로 승격 로직 위치 변경 #21
